requirements.txt
# Core math / utils
numpy>=1.23,<1.27
psutil>=5.9

# OpenCV (Jetson already has system OpenCV; this is optional)
opencv-python>=4.8.0.74

# YOLO / vision
ultralytics>=8.0.0

# PyTorch + torchvision are provided by NVIDIA JetPack on Jetson Orin Nano.
# DO NOT install them from PyPI; follow NVIDIA's wheel instructions instead.

# Audio / alerts
playsound==1.3.0
pyaudio

# GUI and dashboard
pyside6
pygame

# For future hardware / GPIO integrations on Jetson (safe to keep even if unused now)
Jetson.GPIO
pyserial